fix: update stock validation to accept integer input
- Changed stock validation from double to integer parsing to ensure correct input type for stock values.
- Improved error messaging for invalid stock entries to enhance user experience.

+++ lib/app/modules/MarketPlace/Partner/Product/AddProduct/views/add_product_view.dart
@@ -555,7 +555,7 @@ class AddProductView extends GetView<AddProductController> {
                                             if (val == null || val.isEmpty) {
                                               return "Please enter stock";
                                             }
-                                            if (double.tryParse(val) == null) {
+                                            if (int.tryParse(val) == null) {
                                               return "Enter valid number";
                                             }
                                             if (int.tryParse(val) == 0) {
